Update README.md
I designed and executed the "Number Guessing Game" using HTML/CSS to deliver an engaging and interactive gaming experience. This project challenges players to test their intuition and numerical skills in a dynamic virtual environment. With a sleek and intuitive user interface, it offers hours of entertainment while showcasing proficiency in web development.